backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
openpyxl==3.1.5
python-multipart==0.0.9
cryptography==43.0.1
